Wrote unit test for complex merge; changed merge behaviour
Merging two datasets now also merges their columns, based on the type of join which was specified.
Merging records now requires a dataset and initialises new record based on the columns of that dataset.
Attempting to merge without a left, right, or inner join now returns a blank dataset instead of undefined behaviour.